Clean up optimization code path
[x] Move dual averaging and lbfgs into optimizers
[ ] Clean up lbfgs (use scan instead of while_loop
[ ] Refactor pathfinder init to use scan
