+++ blackjax/adaptation/step_size.py
@@ -4,7 +4,7 @@ from typing import Callable, NamedTuple, Tuple
 import jax
 import jax.numpy as jnp
 
-import blackjax.adaptation.optimizers as optimizers
+import blackjax.optimizers.dual_averaging as dual_averaging
 from blackjax.mcmc.hmc import HMCState
 
 __all__ = [
@@ -108,7 +108,7 @@ def dual_averaging_adaptation(
            adaptively setting path lengths in Hamiltonian Monte Carlo." Journal
            of Machine Learning Research 15.1 (2014): 1593-1623.
     """
-    da_init, da_update, da_final = optimizers.dual_averaging(t0, gamma, kappa)
+    da_init, da_update, da_final = dual_averaging.dual_averaging(t0, gamma, kappa)
 
     def init(inital_step_size: float) -> DualAveragingAdaptationState:
         """Initialize the state of the dual averaging scheme.
