+++ blackjax/optimizers/dual_averaging.py
